Ingest .sav files from within a .zip file

This is a common format for microdato DANE. Why bother?, you could simply call haven::read_sav passing the path to the zip file and haven is smart enough to decompress it for you. Well, yeah, in theory. But there are a few glitches that end up breaking that approach. First, Microdato Dane sometimes bundles several different files in a single .zip file (e.g. a .sav, along plain-text -csv, txt- and sometimes also .dta files). This creates trouble for haven (haven usually can decompress the file and warn you there are several files, but frequently end up failing). In addition, sometimes the name of the .sav file does not match that of the .zip file, also creating issues. Finally, in a few cases, several .sav files are bundled in the .zip file, and this function allows you to ingest them all at once.

ingest_zipped_sav(zipfile_path, ...)

Arguments

zipfile_path

path to the .zip file

...

further arguments passed to haven::read_sav

Value

a tibble with columns

sav_file (character): path of the sav file within the zip file

size (numeric): size of the uncompressed sav file

date (of class "POSIXct"): date of the sav file

sav_data (tibble): the data in sav_file, ingested via haven::read_sav

Details

So this function unzip the .sav files contained in the zip