Join DIVIPOLA municipality variables into lhs

Left join lhs and DIVIPOLA at the municipality level. Useful to quickly bring basic data from municipalities (name, department, long, lat) to a dataset that contains only municipality code.

join_divipola_muni_cod(
  lhs,
  by,
  muni_vars = c("dept_cod", "centro_poblado_cod", "dept_nom", "muni_nom",
    "centro_poblado_nom", "tipo_centro_poblado", "long", "lat", "distrito", "tipo_muni",
    "area_metropolitana", "region_nom", "region_cod", "dept_nom_corto", "provincia",
    "muni_nom_corto")
)

Arguments

lhs

A data.frame-like

by

A character vector indicating the column in `lhs` that contains municipality code

muni_vars

A character vector to customize which variables from DIVIPOLA to bring into `lhs`

Value

lhs with additional DIVIPOLA variables

Examples

df <- data.frame(cod = c("05001", "25001"))
join_divipola_muni_cod(df, by = "cod")
#> ✔ 2 rows [100.0%] successfully joined
#>     cod muni_cod dept_cod centro_poblado_cod     dept_nom     muni_nom
#> 1 05001    05001       05           05001000    ANTIOQUIA     MEDELLÍN
#> 2 25001    25001       25           25001000 CUNDINAMARCA AGUA DE DIOS
#>   centro_poblado_nom     tipo_centro_poblado      long      lat distrito
#> 1           MEDELLÍN CABECERA MUNICIPAL (CM) -75.57600 6.248586     <NA>
#> 2       AGUA DE DIOS CABECERA MUNICIPAL (CM) -74.66922 4.375315     <NA>
#>   tipo_muni                     area_metropolitana region_nom region_cod
#> 1 MUNICIPIO AREA METROPOLITANA DEL VALLE DE ABURRÁ    Central          4
#> 2 MUNICIPIO                                   <NA>    Central          4
#>   dept_nom_corto        provincia muni_nom_corto
#> 1      Antioquia VALLE DEL ABURRA       Medellín
#> 2   Cundinamarca   ALTO MAGDALENA   Agua De Dios